requirements.txt
# Core dependencies
requests>=2.31.0,<3.0.0
urllib3>=2.0.0,<3.0.0
PyYAML>=6.0.1,<7.0.0
python-dotenv>=1.0.0,<2.0.0

# Data analysis
numpy>=1.26.0,<2.0.0
scipy>=1.12.0,<2.0.0
pandas>=2.2.0,<3.0.0

# Dashboard
streamlit>=1.31.0,<2.0.0
plotly>=5.18.0,<6.0.0

# Testing (dev)
# pytest>=8.0.0,<9.0.0
# pytest-cov>=4.1.0,<6.0.0
